# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type ParseSettingsChunkMode = typeof ParseSettingsChunkMode[keyof typeof ParseSettingsChunkMode];


export const ParseSettingsChunkMode = {
  off: 'off',
  '1x1': '1x1',
  '2x2': '2x2',
  '3x3': '3x3',
  '4x4': '4x4',
} as const;

export type ParseSettingsEntityMode = typeof ParseSettingsEntityMode[keyof typeof ParseSettingsEntityMode];


export const ParseSettingsEntityMode = {
  off: 'off',
  nbt: 'nbt',
  eggs: 'eggs',
} as const;

export interface ParseSettings {
  maxCoordsPerPart?: number;
  maxCharsPerPart?: number;
  chunkMode?: ParseSettingsChunkMode;
  entityMode?: ParseSettingsEntityMode;
  blockEntityMode?: boolean;
  biomeMode?: boolean;
}

export type UploadInputChunkMode = typeof UploadInputChunkMode[keyof typeof UploadInputChunkMode];


export const UploadInputChunkMode = {
  off: 'off',
  '1x1': '1x1',
  '2x2': '2x2',
  '3x3': '3x3',
  '4x4': '4x4',
} as const;

export type UploadInputEntityMode = typeof UploadInputEntityMode[keyof typeof UploadInputEntityMode];


export const UploadInputEntityMode = {
  off: 'off',
  nbt: 'nbt',
  eggs: 'eggs',
} as const;

export interface UploadInput {
  file: Blob;
  sessionId?: string;
  maxCoordsPerPart?: number;
  maxCharsPerPart?: number;
  chunkMode?: UploadInputChunkMode;
  entityMode?: UploadInputEntityMode;
  blockEntityMode?: boolean;
  biomeMode?: boolean;
}

export interface UploadResult {
  key: string;
  name: string;
  partCount: number;
  sessionId: string;
  blockCount?: number;
  entityCount?: number;
  blockEntityCount?: number;
  regionCount?: number;
}

export interface LitematicFile {
  key: string;
  name: string;
  partCount: number;
  sessionId: string;
  createdAt: string;
  sizeBytes: number;
  blockCount?: number;
  entityCount?: number;
  blockEntityCount?: number;
  regionCount?: number;
  settings?: ParseSettings;
}

export interface FileInfo {
  key: string;
  name: string;
  partCount: number;
}

export interface FilePart {
  key: string;
  number: number;
  total: number;
  /** Minified JSON string of the part payload */
  data: string;
}

export type ListFilesParams = {
sessionId?: string;
};