# lib/api-zod/src/generated/types/fileInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface FileInfo {
  key: string;
  name: string;
  partCount: number;
}