# lib/api-zod/src/generated/types/filePart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface FilePart {
  key: string;
  number: number;
  total: number;
  /** Minified JSON string of the part payload */
  data: string;
}