# lib/api-zod/src/generated/types/litematicFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ParseSettings } from './parseSettings';

export interface LitematicFile {
  key: string;
  name: string;
  partCount: number;
  sessionId: string;
  createdAt: Date;
  sizeBytes: number;
  blockCount?: number;
  entityCount?: number;
  blockEntityCount?: number;
  regionCount?: number;
  settings?: ParseSettings;
}