# lib/api-zod/src/generated/types/parseSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ParseSettingsChunkMode } from './parseSettingsChunkMode';
import type { ParseSettingsEntityMode } from './parseSettingsEntityMode';

export interface ParseSettings {
  maxCoordsPerPart?: number;
  maxCharsPerPart?: number;
  chunkMode?: ParseSettingsChunkMode;
  entityMode?: ParseSettingsEntityMode;
  blockEntityMode?: boolean;
  biomeMode?: boolean;
}