# lib/api-zod/src/generated/types/uploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadInputChunkMode } from './uploadInputChunkMode';
import type { UploadInputEntityMode } from './uploadInputEntityMode';

export interface UploadInput {
  file: Blob;
  sessionId?: string;
  maxCoordsPerPart?: number;
  maxCharsPerPart?: number;
  chunkMode?: UploadInputChunkMode;
  entityMode?: UploadInputEntityMode;
  blockEntityMode?: boolean;
  biomeMode?: boolean;
}