# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Litematic Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  key: string;
  name: string;
  partCount: number;
  sessionId: string;
  blockCount?: number;
  entityCount?: number;
  blockEntityCount?: number;
  regionCount?: number;
}